+++ docs/src/Fulma/Introduction/View.fs
@@ -33,9 +33,9 @@ open Fulma
 ## Architecture
 
 Fulma has been designed to provide the best experience over the Bulma CSS framework.
-To archieve this goal, we assume the user to follow some conventions.
+To archieve this goal, Fulma prevents you to open lower module of the hierarchie.
 
-Always open the "global" module and not the lower module of the hierarchie. For example, if you want to use the Button element you should follow this code:
+For example, if you want to use the Button element you will need to use `open Fulma.Elements` and not `open Fulma.Elements.Button`.
 
 ```fsharp
 open Fulma.Elements