+++ src/Fulma.Extensions/Tooltip.fs
@@ -4,6 +4,7 @@ open Fulma.BulmaClasses
 open Fulma.Common
 open Fable.Helpers.React.Props
 
+[<RequireQualifiedAccess>]
 module Tooltip =
 
     let [<Literal>] ClassName = "tooltip"