+++ src/Fulma/Components/Breadcrumb.fs
@@ -8,6 +8,7 @@ open Fable.Helpers.React
 open Fable.Helpers.React.Props
 open Fable.Import
 
+[<RequireQualifiedAccess>]
 module Breadcrumb =
 
     module Types =