+++ src/Fulma/Components/Modal.fs
@@ -9,6 +9,7 @@ open Fable.Helpers.React
 open Fable.Helpers.React.Props
 open Fable.Import
 
+[<RequireQualifiedAccess>]
 module Modal =
 
     module Types =