+++ src/Fulma/Elements/Image.fs
@@ -5,6 +5,7 @@ open Fulma.Common
 open Fable.Helpers.React
 open Fable.Helpers.React.Props
 
+[<RequireQualifiedAccess>]
 module Image =
     module Types =
         type IImageSize =