+++ src/Fulma/Elements/Table.fs
@@ -5,6 +5,7 @@ open Fulma.Common
 open Fable.Helpers.React
 open Fable.Helpers.React.Props
 
+[<RequireQualifiedAccess>]
 module Table =
     module Types =
         type TableOption =